hdsentinel-mqtt/requirements.txt
paho-mqtt
blkinfo